第55轮全工作区回归健康认证+clippy清零: cargo test --workspace全绿零失败(dsh-cli 260/canvas-shell 22/wasmrt m32-m44群全过),node core 35/35,verify-diff ALL SCENARIOS PASS; clippy从存量30行清到双crate零警告=死代码治理(测试夹具cfg(test)化overlap/card/theme_ns_view/by_key,废弃nodes helper删除)+语义安全自动修复11处(mut冗余/借用冗余/clippy --fix先审diff后纳)+range-loop重写4处(读写k索引改切片迭代/区间max一行式,22测试钉死语义)+doc注释在include!宏前无效降格//; 清理后release+opt重造938129字节内嵌件并重走真路由全量审计=T2~T8全绿零console(行为零变化实证),HANDOFF账本更新第55轮认证基线; serve=term-51活体; 剩余仍=用户两项拍板(S6d切默认+旧前端下线)

+++ canvas-shell/src/interop.rs
@@ -2,7 +2,7 @@
 //! 纪律：业务逻辑零在此层——只搬运（可证逻辑全在 canvas_shell lib）。
 
 use canvas_shell::values::rpc_envelope;
-use js_sys::{JSON, Reflect};
+use js_sys::JSON;
 use serde_json::{json, Value};
 use std::cell::Cell;
 use std::rc::Rc;
@@ -229,7 +229,7 @@ pub fn observe_bump<F: FnMut() + 'static>(f: F) {
             w.clear_timeout_with_handle(handle.get());
         }
         if let Ok(id) = w.set_timeout_with_callback_and_timeout_and_arguments_0(&fire, 200) {
-            handle.set(id as i32);
+            handle.set(id);
         }
     });
     if let Ok(ro) = ResizeObserver::new(&outer) {
